Small integration test for meteogram (#232)
This PR implements minor changes to the existing meteogram integration
test.

### Summary of changes
* rename test and config to meteogram_small
* change date to 2024-08-01 (official start of ICON archive)
* additional parameters to cover all use cases (derived, aggregated and
instantaneous)
