+++ tests/integration/configs/meteogram_small.yaml
@@ -6,13 +6,13 @@ description: |
 config_label: meteogram-test
 
 dates:
-  - 2024-01-01T00:00
+  - 2024-08-01T00:00
 
 runs:
   - forecaster:
       checkpoint: https://servicedepl.meteoswiss.ch/mlstore#/experiments/409/runs/b30acf68520a4bbd8324c44666561696
       label: stage_C_icon_1km
-      steps: 0/6/6
+      steps: 0/12/6
       config: resources/inference/configs/sgm-forecaster-global-ich1.yaml
       extra_requirements:
         - earthkit-utils<0.2.0
@@ -43,12 +43,16 @@ showcase:
   params:
     - T_2M
     - SP_10M
+    - TOT_PREC6
   meteograms:
     enabled: true
     stations:
       - GVE
+      - SAE
   animations:
     enabled: false
+
+lapse_rate_correction: false
 
 locations:
   output_root: output/
