--- tests/integration/test_experiment.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.mark.longtest
-def test_experiment():
-    pass
